Refactor of the GitHubApiCalls package location.
