+++ public/version.json
@@ -1,4 +1,4 @@
-{"version":"1.0.0-7c2b8b3",
-"git-sha":"7c2b8b307f71b2860bcc9c0bc7bc5a3210fa72fd",
-"timestamp":"2020-11-09T14:35:53-0300"}
+{"version":"1.0.0-24a938f",
+"git-sha":"24a938faa629e488bb94b97a329c82bcc2e910da",
+"timestamp":"2020-11-10T06:00:14-0300"}
